+++ src/Press/Http/Composers/SystemMenuComposer.php
@@ -6,6 +6,7 @@ namespace Orchid\Press\Http\Composers;
 
 use Orchid\Platform\ItemMenu;
 use Orchid\Platform\Dashboard;
+use Orchid\Platform\Menu;
 
 class SystemMenuComposer
 {
@@ -30,7 +31,7 @@ class SystemMenuComposer
     public function compose()
     {
         $this->dashboard->menu
-            ->add('Systems',
+            ->add(Menu::SYSTEMS,
                 ItemMenu::setLabel(__('Content management'))
                     ->setSlug('CMS')
                     ->setIcon('icon-layers')